+++ enmapbox/apps/SpecDeepMap/processing_algorithm_tester.py
@@ -6,7 +6,7 @@ from qgis.core import (QgsProcessingAlgorithm,
                        QgsProcessingParameterBoolean,
                        QgsProcessingParameterFileDestination)
 
-from enmapbox.apps.SpecDeepMap.core_tester_remap_classes import process_images_from_csv
+from enmapbox.apps.SpecDeepMap.core_tester import process_images_from_csv
 
 
 class DL_Tester(QgsProcessingAlgorithm):
